+++ src/revisionGraph/model/commitGraphTypes.ts
@@ -28,6 +28,7 @@ export interface RevisionGraphProjectionOptions {
   readonly showTags: boolean;
   readonly showRemoteBranches: boolean;
   readonly showStashes: boolean;
+  readonly showMergeCommits: boolean;
   readonly showCurrentBranchDescendants: boolean;
 }