+++ test/revisionGraphMessageDispatcher.test.ts
@@ -83,6 +83,7 @@ function createReadyRevisionGraphState(): RevisionGraphViewState {
       showTags: true,
       showRemoteBranches: true,
       showStashes: true,
+      showMergeCommits: false,
       showCurrentBranchDescendants: true
     },
     mergeBlockedTargets: [],